- Shrink ipl_cookie_t from 32-bits to 16-bits (large enough to hold   an SR value or an IPL_* constant). - Take advange of the smaller ipl_cookie_t to shrink kmutex_t from   16 bytes to 8 bytes by overlapping storage where possible. - Implement a RAS-based _lock_cas() for mc68010 systems (Sun2).  See   sun68k/sun68k/isr.c.
Tested on various m68k platforms, but NOT Sun2.  In any case, at least
Sun2 compiles now.

+++ sys/arch/sun68k/include/intr.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: intr.h,v 1.12 2007/02/16 02:53:51 ad Exp $	*/
+/*	$NetBSD: intr.h,v 1.13 2007/03/11 05:22:25 thorpej Exp $	*/
 
 /*
  * Copyright (c) 2001 Matt Fredette.
@@ -71,7 +71,7 @@
 
 typedef int ipl_t;
 typedef struct {
-	int _psl;
+	uint16_t _psl;
 } ipl_cookie_t;
 
 ipl_cookie_t makeiplcookie(ipl_t ipl);
